Change to use non-deprecated methods for registering FCM push token
