+++ fcm/src/main/AndroidManifest.xml
@@ -1,17 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.parse.fcm">
-
-    <application>
-
-        <service
-            android:name=".ParseFirebaseJobService"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="com.firebase.jobdispatcher.ACTION_EXECUTE" />
-            </intent-filter>
-        </service>
-
-    </application>
-
-
-</manifest>
+<manifest package="com.parse.fcm"/>
