+++ gcm/build.gradle
@@ -33,7 +33,7 @@ android {
 dependencies {
     // last version for GCM to be supported
     api "com.google.android.gms:play-services-gcm:12.0.1"
-    api "com.firebase:firebase-jobdispatcher:$firebaseJobdispatcherVersion"
+    api "com.firebase:firebase-jobdispatcher:0.8.5"
     implementation project(':parse')
 }
